requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
huggingface-hub==0.20.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pytest==7.4.3
pytest-cov==4.1.0